FDB dump HT PoC
Use hashtable iteration instead of the linked list to dump FDB entries.
Besides the obvious issues such as using a single static iterator, this
does not handle cleanup in case a dump is interrupted half-way through.
Luckily it's not really advantageous compared to the RB tree way and so
we don't have to deal with it.